Update tests to run properly
